front/checkout: Move classes to private submodules
Make the `__init__.py` a bit cleaner by moving the code to private
submodules.  This also allows importing individual parts from another
submodules in the package, which otherwise is not possible.

Refs SHOOP-2502
